requirements.txt
blinker==1.6.3
click==8.1.3
Flask==2.3.0
Flask-Login==0.6.2
Flask-Minify==0.42
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
htmlmin==0.1.12
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lesscpy==0.15.1
MarkupSafe==2.1.3
packaging==23.2
ply==3.11
psycopg2==2.9.9
python-dotenv==1.0.0
rcssmin==1.1.2
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==2.3.0
xxhash==3.4.1
